server/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
google-generativeai==0.8.4
elevenlabs==1.50.0
python-multipart==0.0.20
PyPDF2==3.0.1
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==11.1.0
